+++ schema/random.go
@@ -46,7 +46,7 @@ func GetRandomCreateVideoOpts() *CreateVideoOpts {
 		SRCWidth:         uint(faker.RandomInt(360, 1080)),
 		SRCHeight:        uint(faker.RandomInt(360, 1080)),
 		Duration:         float32(faker.RandomInt(1, 1000)),
-		Framerate:        uint(faker.RandomInt(20, 60)),
+		Framerate:        float32(faker.RandomInt(20, 60)),
 		IsPortrait:       false,
 		CloudFrontURL:    faker.Letterify("??????????.cloudfront.net"),
 		ProcessedAt:      time.Now().UTC(),
